requirements.txt
keyboard==0.13.5
mss==10.0.0
pillow==11.3.0
pyperclip==1.9.0
pywin32==311
